pyproject.toml
[project]
name = "zitan-resto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "python-fasthtml>=0.12.41",
    "supabase>=2.27.3",
]
